addbrush {DescribeDisplay}	R Documentation

Add brush to plot This adds a rectangle to a ggplot plot indicating the brush position.

Description

Add brush to plot This adds a rectangle to a ggplot plot indicating the brush position.

Usage

addbrush(plot, x, y, width = 0.5, height = 0.5, just = c("left", "top"),
  fill = "transparent", col = "black")

Arguments

plot
plot object

x
x position of brush

y
y position of brush

width
width of brush

height
height of brush

just
which corner of brush should be determined by x and y position

fill
fill colour for brush. Use ggplot-alpha for alpha blending.

col
outline colour of brush

Author(s)

Hadley Wickham h.wickham@gmail.com

[Package DescribeDisplay version 0.2.7 Index]