compact_pcp {DescribeDisplay}	R Documentation

Compact pcp data A parallel coordinates is written out as a series of 1D dotplots. This function compacts it back into one dataset.

Description

Compact pcp data A parallel coordinates is written out as a series of 1D dotplots. This function compacts it back into one dataset.

Usage

compact_pcp(data)

Arguments

data
data to pull points from

Author(s)

Hadley Wickham h.wickham@gmail.com

[Package DescribeDisplay version 0.2.7 Index]