compact_scatmat {DescribeDisplay}	R Documentation

Compact scatmat data A scatter plot matrix is made from the information from the 1Dplots. This function collects only that data.

Description

Compact scatmat data A scatter plot matrix is made from the information from the 1Dplots. This function collects only that data.

Usage

compact_scatmat(data)

Arguments

data
data to display

Author(s)

Barret Schloerke schloerke@gmail.com

[Package DescribeDisplay version 0.2.7 Index]