compact_timeseries {DescribeDisplay}	R Documentation

Compact time series data A time series plot is made from scatterplots with a common x axis. This function pulls the correct information out of the data.

Description

Compact time series data A time series plot is made from scatterplots with a common x axis. This function pulls the correct information out of the data.

Usage

compact_timeseries(data)

Arguments

data
data to display

Author(s)

Barret Schloerke schloerke@gmail.com

[Package DescribeDisplay version 0.2.7 Index]