dd_edges {DescribeDisplay}	R Documentation

Describe display edge data Retrieves the describe display edge data for the given plot number.

Description

Describe display edge data Retrieves the describe display edge data for the given plot number.

Usage

dd_edges(dd, n = 1)

Arguments

dd
list of values from describe display

n
plot number, defaults to first plot

Value

data frame suitable for plotting

Author(s)

Hadley Wickham h.wickham@gmail.com

[Package DescribeDisplay version 0.2.7 Index]