dd_load {DescribeDisplay}	R Documentation

Load describe display Retrieve output of from describe display plugin

Description

Also performs some conversion of data structures to more conveient form so that other functions do not have to repeatedly recompute. Some of these conversions could probably be moved into the Describe Display plugin, but it may be easier to just do them on the R side..

Usage

dd_load(path)

Arguments

path
file path

Value

object of class dd

Author(s)

Hadley Wickham h.wickham@gmail.com

See Also

dd_example for an easier way of loading example files

[Package DescribeDisplay version 0.2.7 Index]