dd_tour_axes {DescribeDisplay}	R Documentation

Describe display tour axis Return representation of axes for specified plot

Description

Describe display tour axis Return representation of axes for specified plot

Usage

dd_tour_axes(plot)

Arguments

plot
list of information of a plot

Author(s)

Hadley Wickham h.wickham [at] gmail.com

[Package DescribeDisplay version 0.2.7 Index]