ggplot.barplot {DescribeDisplay}	R Documentation

Create a nice plot for Bar Plots Create a nice looking plot complete with axes using ggplot.

Description

Create a nice plot for Bar Plots Create a nice looking plot complete with axes using ggplot.

Usage

## S3 method for class 'barplot'
ggplot(data, spine = FALSE, ...)

Arguments

data
plot to display, object created by dd_load()

spine
(not implemented currently) whether to display the barchart as a spine plot

...
arguments passed through to the ggplot function

Author(s)

Barret Schloerke schloerke@gmail.com

Examples

library(ggplot2)
print(ggplot(dd_example("barchart")))

[Package DescribeDisplay version 0.2.7 Index]