ggplot.parcoords {DescribeDisplay}	R Documentation

Create a nice plot for parallel coordinates plot Create a nice looking plot complete with axes using ggplot.

Description

Create a nice plot for parallel coordinates plot Create a nice looking plot complete with axes using ggplot.

Usage

## S3 method for class 'parcoords'
ggplot(data, absoluteX = FALSE, absoluteY = FALSE,
  edges = TRUE, ...)

Arguments

data
plot to display

absoluteX
make the sections proportional horizontally to eachother

absoluteY
make the sections proportional vertically to eachother

edges
boolean value to print edges. Defaults to TRUE.

...
arguments passed to the grob function

Author(s)

Barret Schloerke schloerke@gmail.com

Examples

library(ggplot2)
print(ggplot(dd_example("parcoord")))

[Package DescribeDisplay version 0.2.7 Index]