ggplot.scatmat {DescribeDisplay}	R Documentation

Create a nice plots in a scatter plot matrix Create a nice looking plots in a matrix. The 1d sections along the diagonal have a smooth density while the values are compared to eachother within the matrix.

Description

Create a nice plots in a scatter plot matrix Create a nice looking plots in a matrix. The 1d sections along the diagonal have a smooth density while the values are compared to eachother within the matrix.

Usage

## S3 method for class 'scatmat'
ggplot(data, ...)

Arguments

data
data to display

...
(currently) unused arguments

Author(s)

Barret Schloerke schloerke@gmail.com

Examples

library(ggplot2)
print(ggplot(dd_example("scatmat")))

[Package DescribeDisplay version 0.2.7 Index]