ggplot.timeseries {DescribeDisplay}	R Documentation

Create nice plots for a time series Create nice looking plots complete with axes using ggplot. Produces graphics with a uniform x axis.

Description

Create nice plots for a time series Create nice looking plots complete with axes using ggplot. Produces graphics with a uniform x axis.

Usage

## S3 method for class 'timeseries'
ggplot(data, edges = FALSE, ...)

Arguments

data
to display

edges
Boolean operator to tell whether to try to force the edges or not. Will not work to remove the edges.

...
(currently) unused arguments

Author(s)

Barret Schloerke schloerke@gmail.com

Examples

library(ggplot2)
print(ggplot(dd_example("timeseries")))
print(ggplot(dd_example("timeseries"), edges = TRUE))

[Package DescribeDisplay version 0.2.7 Index]