plot.dd {DescribeDisplay}	R Documentation

Draw dd plot Draw a complete describe display.

Description

If you want to layout multiple dd plots on the same page, you can use grid.layout. If you need even more control, set draw = FALSE and then grid.draw the resulting grob yourself.

Usage

## S3 method for class 'dd'
plot(x, ..., draw = TRUE, axislocation = c(0.1, 0.1),
  size = 0.9, axisgp = gpar(col = "black"), background.color = "grey90")

Arguments

x
dd object to plot

...
(unused)

draw
draw plot, or just return grob

axislocation
location of axes (as x and y position in npc coordinates, ie. between 0 and 1)

size
size of plot as a proportion of the total display area (set to 1 for printed out)

axisgp
color of the axis

background.color
color of in the background of the plot

Details

This function reads a number of options directly out of the descripedisplay datastructure. See the examples for ways to use these.

Value

frame grob containing all panels, note that this does not contain the title or border

Author(s)

Hadley Wickham h.wickham@gmail.com

Examples

plot(dd_example("xyplot"))
plot(dd_example("tour1d"))
plot(dd_example("tour2d"))

[Package DescribeDisplay version 0.2.7 Index]