plot.ddplot {DescribeDisplay}	R Documentation

Plot a dd plot Convenient method to draw a single panel.

Description

This is mainly used for bug testing so that you can pull out a single panel quickly and easily.

Usage

## S3 method for class 'ddplot'
plot(x, ..., axislocation = c(0.1, 0.1), axis.gp = gpar(col
  = "black"), background.color = "grey90")

Arguments

x
object to plot

...
(not used)

axislocation
location of axes (as x and y position in npc coordinates, ie. between 0 and 1)

axis.gp
frame grob containing all panels, note that this does not contain the title or border

background.color
color of in the background of the plot

Author(s)

Hadley Wickham h.wickham@gmail.com

Examples

scatmat <- dd_example("scatmat")
plot(scatmat)
plot(scatmat$plots[[1]])
plot(scatmat$plots[[3]])
plot(scatmat$plots[[4]])

[Package DescribeDisplay version 0.2.7 Index]