removehiddens {DescribeDisplay}	R Documentation

Remove hidden points Will remove all hidden points from the plot.

Description

Remove hidden points Will remove all hidden points from the plot.

Usage

removehiddens(d)

Arguments

d
ddplot object

Author(s)

Hadley Wickham h.wickham@gmail.com

[Package DescribeDisplay version 0.2.7 Index]